+++ src/Models/Service.ts
@@ -28,6 +28,8 @@ export default class Service implements ServiceInterface {
   static readonly ORIGIN_ADMIN = 'admin'
   static readonly ORIGIN_BOT = 'bot'
   static readonly ORIGIN_TEST = 'test'
+  static readonly ORIGIN_DRIVER = 'driver'
+  static readonly WP_CLIENT_ID_DRIVER_APP = 'driver-app'
   static readonly EVENT_CANCEL = 'cancel-service'
   static readonly EVENT_TERMINATE = 'end-service'
   static readonly EVENT_ASSIGN = 'assign-service'